9th_session_ResNet50_on_ImageNet/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
PyYAML>=5.4.1
tensorboard>=2.5.0
pillow>=8.2.0 